+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - Improve performance of appending events under normal and degraded network conditions ([#230](https://github.com/commanded/eventstore/pull/230)).
+- Subscription checkpoint tuning ([#237](https://github.com/commanded/eventstore/pull/237)).
 
 ## v1.2.3